contract-tests/tests/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "contract-tests"
description = "Contract tests for AWS OTEL NodeJS Instrumentation"
version = "1.0.0"
license = "Apache-2.0"
requires-python = ">=3.8"

dependencies = [
    "opentelemetry-proto==1.25.0",
    "opentelemetry-sdk==1.25.0",
    "testcontainers==3.7.1",
    "grpcio==1.60.0",
    "docker==7.1.0",
    "mock-collector==1.0.0",
    "requests==2.32.2"
]

[project.optional-dependencies]
test = []

[tool.hatch.build.targets.sdist]
include = ["/test"]

[tool.hatch.build.targets.wheel]
packages = ["test/amazon"]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"